ENC: Add parallel forward run
Re establishing the parallel forward run for adjoint because Tapenade seems to
be happy with it. The idea is that having re written the solver in the previous
commit allows Tapenade to clarify the OMP atomic directive.

The parallel forward run is available for each module (snow, hydrological, routing).
Snow and hydrological module are fully parallelized on the whole domain. For the
routing module, new variables are calculated while generating the mesh (npar, ncpar,
cscpar, cpar_to_rowcol and flwpar). This variables allow to parallelize the routing
by independent partition.

Dataset meshs have been updated
Tests have been updated