ENH: Low memory adjoint
This commit is an upgrade of the adjoint model allowing time step
checkpoints to lower the maximum memory usage. By default, in the
forward sweep of the adjoint, at each time, n arrays of shape (nrow, ncol)
where pushed. This arrays were poped at the end of the forward sweep which
result in some case (large domain and time steps) to a too high memory peak.
This commit adds checkpoints in the forward sweep to reduce the memory peak
and pass the variables that must be checkpointed at each time (i.e. fluxes and
states) to a vector of active cell instead of a grid of number of rows and columns.

This upgrade should be pretty efficient for large domain and time step calibration