FIX: Beta parameter in GR percolation
- The beta parameters in GR production in time step dependent. Currently, beta
was assigned to the daily time step value (i.e., 9/4)
This commit adds a beta parameter which is time step dependent. The hydrological
module concerned by this modification are ``gr4``, ``gr5`` and ``loieau``

- This modifications impacted the test. A new baseline has been generated.
The setup of the tests on mutliset estimate has been changed to overcome
warning when dividing by zero.

- The docstring has been updated