ENH: Update the reading of geo-referenced input data
The aim of this commit is to address the problems associated with reading geo-referenced data (atmospheric and physiographic data).
In particular, certain specific cases were not fully handled by smash:
- Different spatial resolution
- Overlap missmatch
- Mesh extent greater than data extent.
This commit improves these 3 aspects by returning a warning if one of these cases occurs on one of the files read.
For the resolution, the file is resampled using nearest neighbour algorithm. For overlap missmatch, the reading window is shifted to overlap on the
nearest cell. And, if the mesh extension is too large, the code no longer returns an error but fills the array with no data in the part whose extent
is not covered by the data.

In addition, when generating a mesh with a bounding box, if the given bounding box does not overlap with the flow directions, it is padded.

Finally, a series of tests on the various possible warnings and some documentation has been added