FIX: Fix Fortran character array
This commit addresses problems with Fortran character arrays passed to Python. The first was that the length of the character string
was too short (20). The second was that only 1D arrays were managed and if the length of the character string is equal to the limit
then the values in the array are not splited correctly.

The max length has been increased from 20 to 128 and the associated decorator to handle character arrays has been modified to managed
any nD arrays and max character length.

A few tests have been added, but they don't allow you to check all the decorators correctly. It would be necessary to create a
Fortran file specific to the test case.